requirements.txt
Flask==2.2.2
pytest==7.2.1
requests==2.28.2

Pillow~=9.4.0
paho-mqtt~=1.6.1